Crate rusoto_storagegateway

AWS Storage Gateway Service
AWS Storage Gateway is the service that connects an on-premises software appliance with cloud-based storage to provide seamless and secure integration between an organization's on-premises IT environment and the AWS storage infrastructure. The service enables you to securely upload data to the AWS cloud for cost effective backup and rapid disaster recovery.

Use the following links to get started using the AWS Storage Gateway Service API Reference:

AWS Storage Gateway Required Request Headers: Describes the required headers that you must send with every POST request to AWS Storage Gateway.

Signing Requests: AWS Storage Gateway requires that you authenticate every request you send; this topic describes how sign such a request.

Error Responses: Provides reference information about AWS Storage Gateway errors.

Operations in AWS Storage Gateway: Contains detailed descriptions of all AWS Storage Gateway operations, their request parameters, response elements, possible errors, and examples of requests and responses.

AWS Storage Gateway Regions and Endpoints: Provides a list of each AWS region and endpoints available for use with AWS Storage Gateway.

AWS Storage Gateway resource IDs are in uppercase. When you use these resource IDs with the Amazon EC2 API, EC2 expects resource IDs in lowercase. You must change your resource ID to lowercase to use it with the EC2 API. For example, in Storage Gateway the ID for a volume might be vol-AA22BB012345DAF670. When you use this ID with the EC2 API, you must change it to vol-aa22bb012345daf670. Otherwise, the EC2 API might not behave as expected.

IDs for Storage Gateway volumes and Amazon EBS snapshots created from gateway volumes are changing to a longer format. Starting in December 2016, all new volumes and snapshots will be created with a 17-character string. Starting in April 2016, you will be able to use these longer IDs so you can test your systems with the new format. For more information, see Longer EC2 and EBS Resource IDs.

For example, a volume Amazon Resource Name (ARN) with the longer volume ID format looks like the following:

arn:aws:storagegateway:us-west-2:111122223333:gateway/sgw-12A3456B/volume/vol-1122AABBCCDDEEFFG.

A snapshot ID with the longer ID format looks like the following: snap-78e226633445566ee.

For more information, see Announcement: Heads-up – Longer AWS Storage Gateway volume and snapshot IDs coming in 2016.

If you're using the service, you're probably looking for StorageGatewayClient and StorageGateway.

Structs

ActivateGatewayInput

A JSON object containing one or more of the following fields:

ActivateGatewayInput$ActivationKey

ActivateGatewayInput$GatewayName

ActivateGatewayInput$GatewayRegion

ActivateGatewayInput$GatewayTimezone

ActivateGatewayInput$GatewayType

ActivateGatewayInput$TapeDriveType

ActivateGatewayInput$MediumChangerType

ActivateGatewayOutput

AWS Storage Gateway returns the Amazon Resource Name (ARN) of the activated gateway. It is a string made of information such as your account, gateway name, and region. This ARN is used to reference the gateway in other API operations as well as resource-based authorization.

For gateways activated prior to September 02, 2015, the gateway ARN contains the gateway name rather than the gateway ID. Changing the name of the gateway has no effect on the gateway ARN.

AddCacheInput

AddCacheOutput

AddTagsToResourceInput
AddTagsToResourceInput

AddTagsToResourceOutput
AddTagsToResourceOutput

AddUploadBufferInput

AddUploadBufferOutput

AddWorkingStorageInput

A JSON object containing one or more of the following fields:

AddWorkingStorageInput$DiskIds

AddWorkingStorageOutput
A JSON object containing the of the gateway for which working storage was configured.

AssignTapePoolInput

AssignTapePoolOutput

AttachVolumeInput
AttachVolumeInput

AttachVolumeOutput
AttachVolumeOutput

CachediSCSIVolume
Describes an iSCSI cached volume.

CancelArchivalInput
CancelArchivalInput

CancelArchivalOutput
CancelArchivalOutput

CancelRetrievalInput
CancelRetrievalInput

CancelRetrievalOutput
CancelRetrievalOutput

ChapInfo
Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

CreateCachediSCSIVolumeInput

CreateCachediSCSIVolumeOutput

CreateNFSFileShareInput
CreateNFSFileShareInput

CreateNFSFileShareOutput
CreateNFSFileShareOutput

CreateSMBFileShareInput
CreateSMBFileShareInput

CreateSMBFileShareOutput
CreateSMBFileShareOutput

CreateSnapshotFromVolumeRecoveryPointInput

CreateSnapshotFromVolumeRecoveryPointOutput

CreateSnapshotInput

A JSON object containing one or more of the following fields:

CreateSnapshotInput$SnapshotDescription

CreateSnapshotInput$VolumeARN

CreateSnapshotOutput
A JSON object containing the following fields:

CreateStorediSCSIVolumeInput

A JSON object containing one or more of the following fields:

CreateStorediSCSIVolumeInput$DiskId

CreateStorediSCSIVolumeInput$NetworkInterfaceId

CreateStorediSCSIVolumeInput$PreserveExistingData

CreateStorediSCSIVolumeInput$SnapshotId

CreateStorediSCSIVolumeInput$TargetName

CreateStorediSCSIVolumeOutput
A JSON object containing the following fields:

CreateTapeWithBarcodeInput
CreateTapeWithBarcodeInput

CreateTapeWithBarcodeOutput
CreateTapeOutput

CreateTapesInput
CreateTapesInput

CreateTapesOutput
CreateTapeOutput

DeleteBandwidthRateLimitInput

A JSON object containing the following fields:

DeleteBandwidthRateLimitInput$BandwidthType

DeleteBandwidthRateLimitOutput
A JSON object containing the of the gateway whose bandwidth rate information was deleted.

DeleteChapCredentialsInput

A JSON object containing one or more of the following fields:

DeleteChapCredentialsInput$InitiatorName

DeleteChapCredentialsInput$TargetARN

DeleteChapCredentialsOutput
A JSON object containing the following fields:

DeleteFileShareInput
DeleteFileShareInput

DeleteFileShareOutput
DeleteFileShareOutput

DeleteGatewayInput
A JSON object containing the ID of the gateway to delete.

DeleteGatewayOutput
A JSON object containing the ID of the deleted gateway.

DeleteSnapshotScheduleInput

DeleteSnapshotScheduleOutput

DeleteTapeArchiveInput
DeleteTapeArchiveInput

DeleteTapeArchiveOutput
DeleteTapeArchiveOutput

DeleteTapeInput
DeleteTapeInput

DeleteTapeOutput
DeleteTapeOutput

DeleteVolumeInput
A JSON object containing the DeleteVolumeInput$VolumeARN to delete.

DeleteVolumeOutput
A JSON object containing the of the storage volume that was deleted

DescribeBandwidthRateLimitInput
A JSON object containing the of the gateway.

DescribeBandwidthRateLimitOutput
A JSON object containing the following fields:

DescribeCacheInput

DescribeCacheOutput

DescribeCachediSCSIVolumesInput

DescribeCachediSCSIVolumesOutput
A JSON object containing the following fields:

DescribeChapCredentialsInput
A JSON object containing the Amazon Resource Name (ARN) of the iSCSI volume target.

DescribeChapCredentialsOutput
A JSON object containing a .

DescribeGatewayInformationInput
A JSON object containing the ID of the gateway.

DescribeGatewayInformationOutput
A JSON object containing the following fields:

DescribeMaintenanceStartTimeInput
A JSON object containing the of the gateway.

DescribeMaintenanceStartTimeOutput

A JSON object containing the following fields:

DescribeMaintenanceStartTimeOutput$DayOfMonth

DescribeMaintenanceStartTimeOutput$DayOfWeek

DescribeMaintenanceStartTimeOutput$HourOfDay

DescribeMaintenanceStartTimeOutput$MinuteOfHour

DescribeMaintenanceStartTimeOutput$Timezone

DescribeNFSFileSharesInput
DescribeNFSFileSharesInput

DescribeNFSFileSharesOutput
DescribeNFSFileSharesOutput

DescribeSMBFileSharesInput
DescribeSMBFileSharesInput

DescribeSMBFileSharesOutput
DescribeSMBFileSharesOutput

DescribeSMBSettingsInput

DescribeSMBSettingsOutput

DescribeSnapshotScheduleInput
A JSON object containing the DescribeSnapshotScheduleInput$VolumeARN of the volume.

DescribeSnapshotScheduleOutput

DescribeStorediSCSIVolumesInput
A JSON object containing a list of DescribeStorediSCSIVolumesInput$VolumeARNs.

DescribeStorediSCSIVolumesOutput

DescribeTapeArchivesInput
DescribeTapeArchivesInput

DescribeTapeArchivesOutput
DescribeTapeArchivesOutput

DescribeTapeRecoveryPointsInput
DescribeTapeRecoveryPointsInput

DescribeTapeRecoveryPointsOutput
DescribeTapeRecoveryPointsOutput

DescribeTapesInput
DescribeTapesInput

DescribeTapesOutput
DescribeTapesOutput

DescribeUploadBufferInput

DescribeUploadBufferOutput

DescribeVTLDevicesInput
DescribeVTLDevicesInput

DescribeVTLDevicesOutput
DescribeVTLDevicesOutput

DescribeWorkingStorageInput
A JSON object containing the of the gateway.

DescribeWorkingStorageOutput
A JSON object containing the following fields:

DetachVolumeInput
AttachVolumeInput

DetachVolumeOutput
AttachVolumeOutput

DeviceiSCSIAttributes
Lists iSCSI information about a VTL device.

DisableGatewayInput
DisableGatewayInput

DisableGatewayOutput
DisableGatewayOutput

Disk
Represents a gateway's local disk.

FileShareInfo
Describes a file share.

GatewayInfo
Describes a gateway object.

JoinDomainInput
JoinDomainInput

JoinDomainOutput
JoinDomainOutput

ListFileSharesInput
ListFileShareInput

ListFileSharesOutput
ListFileShareOutput

ListGatewaysInput

A JSON object containing zero or more of the following fields:

ListGatewaysInput$Limit

ListGatewaysInput$Marker

ListGatewaysOutput

ListLocalDisksInput
A JSON object containing the of the gateway.

ListLocalDisksOutput

ListTagsForResourceInput
ListTagsForResourceInput

ListTagsForResourceOutput
ListTagsForResourceOutput

ListTapesInput

A JSON object that contains one or more of the following fields:

ListTapesInput$Limit

ListTapesInput$Marker

ListTapesInput$TapeARNs

ListTapesOutput

A JSON object containing the following fields:

ListTapesOutput$Marker

ListTapesOutput$VolumeInfos

ListVolumeInitiatorsInput
ListVolumeInitiatorsInput

ListVolumeInitiatorsOutput
ListVolumeInitiatorsOutput

ListVolumeRecoveryPointsInput

ListVolumeRecoveryPointsOutput

ListVolumesInput

A JSON object that contains one or more of the following fields:

ListVolumesInput$Limit

ListVolumesInput$Marker

ListVolumesOutput

A JSON object containing the following fields:

ListVolumesOutput$Marker

ListVolumesOutput$VolumeInfos

NFSFileShareDefaults
Describes Network File System (NFS) file share default values. Files and folders stored as Amazon S3 objects in S3 buckets don't, by default, have Unix file permissions assigned to them. Upon discovery in an S3 bucket by Storage Gateway, the S3 objects that represent files and folders are assigned these default Unix permissions. This operation is only supported for file gateways.

NFSFileShareInfo
The Unix file permissions and ownership information assigned, by default, to native S3 objects when file gateway discovers them in S3 buckets. This operation is only supported in file gateways.

NetworkInterface
Describes a gateway's network interface.

NotifyWhenUploadedInput

NotifyWhenUploadedOutput

RefreshCacheInput
RefreshCacheInput

RefreshCacheOutput
RefreshCacheOutput

RemoveTagsFromResourceInput
RemoveTagsFromResourceInput

RemoveTagsFromResourceOutput
RemoveTagsFromResourceOutput

ResetCacheInput

ResetCacheOutput

RetrieveTapeArchiveInput
RetrieveTapeArchiveInput

RetrieveTapeArchiveOutput
RetrieveTapeArchiveOutput

RetrieveTapeRecoveryPointInput
RetrieveTapeRecoveryPointInput

RetrieveTapeRecoveryPointOutput
RetrieveTapeRecoveryPointOutput

SMBFileShareInfo
The Windows file permissions and ownership information assigned, by default, to native S3 objects when file gateway discovers them in S3 buckets. This operation is only supported for file gateways.

SetLocalConsolePasswordInput
SetLocalConsolePasswordInput

SetLocalConsolePasswordOutput

SetSMBGuestPasswordInput
SetSMBGuestPasswordInput

SetSMBGuestPasswordOutput

ShutdownGatewayInput
A JSON object containing the of the gateway to shut down.

ShutdownGatewayOutput
A JSON object containing the of the gateway that was shut down.

StartGatewayInput
A JSON object containing the of the gateway to start.

StartGatewayOutput
A JSON object containing the of the gateway that was restarted.

StorageGatewayClient
A client for the AWS Storage Gateway API.

StorageGatewayError
Provides additional information about an error that was returned by the service as an or. See the errorCode and errorDetails members for more information about the error.

StorediSCSIVolume
Describes an iSCSI stored volume.

Tag
A key-value pair that helps you manage, filter, and search for your resource. Allowed characters: letters, white space, and numbers, representable in UTF-8, and the following characters: + - = . _ : /

Tape
Describes a virtual tape object.

TapeArchive
Represents a virtual tape that is archived in the virtual tape shelf (VTS).

TapeInfo
Describes a virtual tape.

TapeRecoveryPointInfo
Describes a recovery point.

UpdateBandwidthRateLimitInput

A JSON object containing one or more of the following fields:

UpdateBandwidthRateLimitInput$AverageDownloadRateLimitInBitsPerSec

UpdateBandwidthRateLimitInput$AverageUploadRateLimitInBitsPerSec

UpdateBandwidthRateLimitOutput
A JSON object containing the of the gateway whose throttle information was updated.

UpdateChapCredentialsInput

A JSON object containing one or more of the following fields:

UpdateChapCredentialsInput$InitiatorName

UpdateChapCredentialsInput$SecretToAuthenticateInitiator

UpdateChapCredentialsInput$SecretToAuthenticateTarget

UpdateChapCredentialsInput$TargetARN

UpdateChapCredentialsOutput
A JSON object containing the following fields:

UpdateGatewayInformationInput

UpdateGatewayInformationOutput
A JSON object containing the ARN of the gateway that was updated.

UpdateGatewaySoftwareNowInput
A JSON object containing the of the gateway to update.

UpdateGatewaySoftwareNowOutput
A JSON object containing the of the gateway that was updated.

UpdateMaintenanceStartTimeInput

A JSON object containing the following fields:

UpdateMaintenanceStartTimeInput$DayOfMonth

UpdateMaintenanceStartTimeInput$DayOfWeek

UpdateMaintenanceStartTimeInput$HourOfDay

UpdateMaintenanceStartTimeInput$MinuteOfHour

UpdateMaintenanceStartTimeOutput
A JSON object containing the of the gateway whose maintenance start time is updated.

UpdateNFSFileShareInput
UpdateNFSFileShareInput

UpdateNFSFileShareOutput
UpdateNFSFileShareOutput

UpdateSMBFileShareInput
UpdateSMBFileShareInput

UpdateSMBFileShareOutput
UpdateSMBFileShareOutput

UpdateSMBSecurityStrategyInput

UpdateSMBSecurityStrategyOutput

UpdateSnapshotScheduleInput

A JSON object containing one or more of the following fields:

UpdateSnapshotScheduleInput$Description

UpdateSnapshotScheduleInput$RecurrenceInHours

UpdateSnapshotScheduleInput$StartAt

UpdateSnapshotScheduleInput$VolumeARN

UpdateSnapshotScheduleOutput
A JSON object containing the of the updated storage volume.

UpdateVTLDeviceTypeInput

UpdateVTLDeviceTypeOutput
UpdateVTLDeviceTypeOutput

VTLDevice
Represents a device object associated with a tape gateway.

VolumeInfo
Describes a storage volume object.

VolumeRecoveryPointInfo
Describes a storage volume recovery point object.

VolumeiSCSIAttributes
Lists iSCSI information about a volume.

Enums

ActivateGatewayError
Errors returned by ActivateGateway

AddCacheError
Errors returned by AddCache

AddTagsToResourceError
Errors returned by AddTagsToResource

AddUploadBufferError
Errors returned by AddUploadBuffer

AddWorkingStorageError
Errors returned by AddWorkingStorage

AssignTapePoolError
Errors returned by AssignTapePool

AttachVolumeError
Errors returned by AttachVolume

CancelArchivalError
Errors returned by CancelArchival

CancelRetrievalError
Errors returned by CancelRetrieval

CreateCachediSCSIVolumeError
Errors returned by CreateCachediSCSIVolume

CreateNFSFileShareError
Errors returned by CreateNFSFileShare

CreateSMBFileShareError
Errors returned by CreateSMBFileShare

CreateSnapshotError
Errors returned by CreateSnapshot

CreateSnapshotFromVolumeRecoveryPointError
Errors returned by CreateSnapshotFromVolumeRecoveryPoint

CreateStorediSCSIVolumeError
Errors returned by CreateStorediSCSIVolume

CreateTapeWithBarcodeError
Errors returned by CreateTapeWithBarcode

CreateTapesError
Errors returned by CreateTapes

DeleteBandwidthRateLimitError
Errors returned by DeleteBandwidthRateLimit

DeleteChapCredentialsError
Errors returned by DeleteChapCredentials

DeleteFileShareError
Errors returned by DeleteFileShare

DeleteGatewayError
Errors returned by DeleteGateway

DeleteSnapshotScheduleError
Errors returned by DeleteSnapshotSchedule

DeleteTapeArchiveError
Errors returned by DeleteTapeArchive

DeleteTapeError
Errors returned by DeleteTape

DeleteVolumeError
Errors returned by DeleteVolume

DescribeBandwidthRateLimitError
Errors returned by DescribeBandwidthRateLimit

DescribeCacheError
Errors returned by DescribeCache

DescribeCachediSCSIVolumesError
Errors returned by DescribeCachediSCSIVolumes

DescribeChapCredentialsError
Errors returned by DescribeChapCredentials

DescribeGatewayInformationError
Errors returned by DescribeGatewayInformation

DescribeMaintenanceStartTimeError
Errors returned by DescribeMaintenanceStartTime

DescribeNFSFileSharesError
Errors returned by DescribeNFSFileShares

DescribeSMBFileSharesError
Errors returned by DescribeSMBFileShares

DescribeSMBSettingsError
Errors returned by DescribeSMBSettings

DescribeSnapshotScheduleError
Errors returned by DescribeSnapshotSchedule

DescribeStorediSCSIVolumesError
Errors returned by DescribeStorediSCSIVolumes

DescribeTapeArchivesError
Errors returned by DescribeTapeArchives

DescribeTapeRecoveryPointsError
Errors returned by DescribeTapeRecoveryPoints

DescribeTapesError
Errors returned by DescribeTapes

DescribeUploadBufferError
Errors returned by DescribeUploadBuffer

DescribeVTLDevicesError
Errors returned by DescribeVTLDevices

DescribeWorkingStorageError
Errors returned by DescribeWorkingStorage

DetachVolumeError
Errors returned by DetachVolume

DisableGatewayError
Errors returned by DisableGateway

JoinDomainError
Errors returned by JoinDomain

ListFileSharesError
Errors returned by ListFileShares

ListGatewaysError
Errors returned by ListGateways

ListLocalDisksError
Errors returned by ListLocalDisks

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTapesError
Errors returned by ListTapes

ListVolumeInitiatorsError
Errors returned by ListVolumeInitiators

ListVolumeRecoveryPointsError
Errors returned by ListVolumeRecoveryPoints

ListVolumesError
Errors returned by ListVolumes

NotifyWhenUploadedError
Errors returned by NotifyWhenUploaded

RefreshCacheError
Errors returned by RefreshCache

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

ResetCacheError
Errors returned by ResetCache

RetrieveTapeArchiveError
Errors returned by RetrieveTapeArchive

RetrieveTapeRecoveryPointError
Errors returned by RetrieveTapeRecoveryPoint

SetLocalConsolePasswordError
Errors returned by SetLocalConsolePassword

SetSMBGuestPasswordError
Errors returned by SetSMBGuestPassword

ShutdownGatewayError
Errors returned by ShutdownGateway

StartGatewayError
Errors returned by StartGateway

UpdateBandwidthRateLimitError
Errors returned by UpdateBandwidthRateLimit

UpdateChapCredentialsError
Errors returned by UpdateChapCredentials

UpdateGatewayInformationError
Errors returned by UpdateGatewayInformation

UpdateGatewaySoftwareNowError
Errors returned by UpdateGatewaySoftwareNow

UpdateMaintenanceStartTimeError
Errors returned by UpdateMaintenanceStartTime

UpdateNFSFileShareError
Errors returned by UpdateNFSFileShare

UpdateSMBFileShareError
Errors returned by UpdateSMBFileShare

UpdateSMBSecurityStrategyError
Errors returned by UpdateSMBSecurityStrategy

UpdateSnapshotScheduleError
Errors returned by UpdateSnapshotSchedule

UpdateVTLDeviceTypeError
Errors returned by UpdateVTLDeviceType

Traits

StorageGateway
Trait representing the capabilities of the AWS Storage Gateway API. AWS Storage Gateway clients implement this trait.