Enum rusoto_storagegateway::RetrieveTapeRecoveryPointError

pub enum RetrieveTapeRecoveryPointError {
    InternalServerError(String),
    InvalidGatewayRequest(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by RetrieveTapeRecoveryPoint

Variants
InternalServerError(String)

An internal server error has occurred during the request. For more information, see the error and message fields.

InvalidGatewayRequest(String)

An exception occurred because an invalid gateway request was issued to the service. For more information, see the error and message fields.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl RetrieveTapeRecoveryPointError

fn from_body(body: &str) -> RetrieveTapeRecoveryPointError