Struct rusoto_storagegateway::CreateStorediSCSIVolumeInput

pub struct CreateStorediSCSIVolumeInput {
    pub disk_id: String,
    pub gateway_arn: String,
    pub network_interface_id: String,
    pub preserve_existing_data: bool,
    pub snapshot_id: Option<String>,
    pub target_name: String,
}

A JSON object containing one or more of the following fields:

CreateStorediSCSIVolumeInput$DiskId

CreateStorediSCSIVolumeInput$NetworkInterfaceId

CreateStorediSCSIVolumeInput$PreserveExistingData

CreateStorediSCSIVolumeInput$SnapshotId

CreateStorediSCSIVolumeInput$TargetName

Fields
disk_id: String

The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

gateway_arn: String network_interface_id: String

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

preserve_existing_data: bool

Specify this field as true if you want to preserve the data on the local disk. Otherwise, specifying this field as false creates an empty volume.

Valid Values: true, false

snapshot_id: Option<String>

The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot otherwise do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

target_name: String

The name of the iSCSI target used by initiators to connect to the target and as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes of a gateway.