Struct rusoto_storagegateway::NFSFileShareInfo

pub struct NFSFileShareInfo {
    pub client_list: Option<Vec<String>>,
    pub default_storage_class: Option<String>,
    pub file_share_arn: Option<String>,
    pub file_share_id: Option<String>,
    pub file_share_status: Option<String>,
    pub gateway_arn: Option<String>,
    pub kms_encrypted: Option<bool>,
    pub kms_key: Option<String>,
    pub location_arn: Option<String>,
    pub nfs_file_share_defaults: Option<NFSFileShareDefaults>,
    pub path: Option<String>,
    pub read_only: Option<bool>,
    pub role: Option<String>,
    pub squash: Option<String>,
}

The Unix file permissions and ownership information assigned, by default, to native S3 objects when file gateway discovers them in S3 buckets. This operation is only supported in file gateways.

Fields
client_list: Option<Vec<String>> default_storage_class: Option<String>

The default storage class for objects put into an Amazon S3 bucket by file gateway. Possible values are S3_STANDARD or S3_STANDARD_IA. If this field is not populated, the default value S3_STANDARD is used. Optional.

file_share_arn: Option<String> file_share_id: Option<String> file_share_status: Option<String> gateway_arn: Option<String> kms_encrypted: Option<bool>

True to use Amazon S3 server side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

kms_key: Option<String> location_arn: Option<String> nfs_file_share_defaults: Option<NFSFileShareDefaults> path: Option<String> read_only: Option<bool> role: Option<String> squash: Option<String>