Struct rusoto_storagegateway::StorageGatewayClient

pub struct StorageGatewayClient { /* fields omitted */ }

A client for the AWS Storage Gateway API.

Methods

impl StorageGatewayClient

pub fn new(region: Region) -> StorageGatewayClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> StorageGatewayClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,