Crate rusoto_storagegateway

AWS Storage Gateway

If you're using the service, you're probably looking for StorageGatewayClient and StorageGateway.

Structs

ActivateGatewayInput

A JSON object containing one or more of the following fields:

ActivateGatewayInput$ActivationKey

ActivateGatewayInput$GatewayName

ActivateGatewayInput$GatewayRegion

ActivateGatewayInput$GatewayTimezone

ActivateGatewayInput$GatewayType

ActivateGatewayInput$TapeDriveType

ActivateGatewayInput$MediumChangerType

ActivateGatewayOutput

AWS Storage Gateway returns the Amazon Resource Name (ARN) of the activated gateway. It is a string made of information such as your account, gateway name, and region. This ARN is used to reference the gateway in other API operations as well as resource-based authorization.

For gateways activated prior to September 02, 2015 the gateway ARN contains the gateway name rather than the gateway id. Changing the name of the gateway has no effect on the gateway ARN.

AddCacheInput

AddCacheOutput

AddTagsToResourceInput
AddTagsToResourceInput

AddTagsToResourceOutput
AddTagsToResourceOutput

AddUploadBufferInput

AddUploadBufferOutput

AddWorkingStorageInput

A JSON object containing one or more of the following fields:

AddWorkingStorageInput$DiskIds

AddWorkingStorageOutput
A JSON object containing the of the gateway for which working storage was configured.

CachediSCSIVolume
Describes an iSCSI cached volume.

CancelArchivalInput
CancelArchivalInput

CancelArchivalOutput
CancelArchivalOutput

CancelRetrievalInput
CancelRetrievalInput

CancelRetrievalOutput
CancelRetrievalOutput

ChapInfo
Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

CreateCachediSCSIVolumeInput

CreateCachediSCSIVolumeOutput

CreateNFSFileShareInput
CreateNFSFileShareInput

CreateNFSFileShareOutput
CreateNFSFileShareOutput

CreateSnapshotFromVolumeRecoveryPointInput

CreateSnapshotFromVolumeRecoveryPointOutput

CreateSnapshotInput

A JSON object containing one or more of the following fields:

CreateSnapshotInput$SnapshotDescription

CreateSnapshotInput$VolumeARN

CreateSnapshotOutput
A JSON object containing the following fields:

CreateStorediSCSIVolumeInput

A JSON object containing one or more of the following fields:

CreateStorediSCSIVolumeInput$DiskId

CreateStorediSCSIVolumeInput$NetworkInterfaceId

CreateStorediSCSIVolumeInput$PreserveExistingData

CreateStorediSCSIVolumeInput$SnapshotId

CreateStorediSCSIVolumeInput$TargetName

CreateStorediSCSIVolumeOutput
A JSON object containing the following fields:

CreateTapeWithBarcodeInput
CreateTapeWithBarcodeInput

CreateTapeWithBarcodeOutput
CreateTapeOutput

CreateTapesInput
CreateTapesInput

CreateTapesOutput
CreateTapeOutput

DeleteBandwidthRateLimitInput

A JSON object containing the following fields:

DeleteBandwidthRateLimitInput$BandwidthType

DeleteBandwidthRateLimitOutput
A JSON object containing the of the gateway whose bandwidth rate information was deleted.

DeleteChapCredentialsInput

A JSON object containing one or more of the following fields:

DeleteChapCredentialsInput$InitiatorName

DeleteChapCredentialsInput$TargetARN

DeleteChapCredentialsOutput
A JSON object containing the following fields:

DeleteFileShareInput
DeleteFileShareInput

DeleteFileShareOutput
DeleteFileShareOutput

DeleteGatewayInput
A JSON object containing the id of the gateway to delete.

DeleteGatewayOutput
A JSON object containing the id of the deleted gateway.

DeleteSnapshotScheduleInput

DeleteSnapshotScheduleOutput

DeleteTapeArchiveInput
DeleteTapeArchiveInput

DeleteTapeArchiveOutput
DeleteTapeArchiveOutput

DeleteTapeInput
DeleteTapeInput

DeleteTapeOutput
DeleteTapeOutput

DeleteVolumeInput
A JSON object containing the DeleteVolumeInput$VolumeARN to delete.

DeleteVolumeOutput
A JSON object containing the of the storage volume that was deleted

DescribeBandwidthRateLimitInput
A JSON object containing the of the gateway.

DescribeBandwidthRateLimitOutput
A JSON object containing the following fields:

DescribeCacheInput

DescribeCacheOutput

DescribeCachediSCSIVolumesInput

DescribeCachediSCSIVolumesOutput
A JSON object containing the following fields:

DescribeChapCredentialsInput
A JSON object containing the Amazon Resource Name (ARN) of the iSCSI volume target.

DescribeChapCredentialsOutput
A JSON object containing a .

DescribeGatewayInformationInput
A JSON object containing the id of the gateway.

DescribeGatewayInformationOutput
A JSON object containing the following fields:

DescribeMaintenanceStartTimeInput
A JSON object containing the of the gateway.

DescribeMaintenanceStartTimeOutput

A JSON object containing the following fields:

DescribeMaintenanceStartTimeOutput$DayOfWeek

DescribeMaintenanceStartTimeOutput$HourOfDay

DescribeMaintenanceStartTimeOutput$MinuteOfHour

DescribeMaintenanceStartTimeOutput$Timezone

DescribeNFSFileSharesInput
DescribeNFSFileSharesInput

DescribeNFSFileSharesOutput
DescribeNFSFileSharesOutput

DescribeSnapshotScheduleInput
A JSON object containing the DescribeSnapshotScheduleInput$VolumeARN of the volume.

DescribeSnapshotScheduleOutput

DescribeStorediSCSIVolumesInput
A JSON object containing a list of DescribeStorediSCSIVolumesInput$VolumeARNs.

DescribeStorediSCSIVolumesOutput

DescribeTapeArchivesInput
DescribeTapeArchivesInput

DescribeTapeArchivesOutput
DescribeTapeArchivesOutput

DescribeTapeRecoveryPointsInput
DescribeTapeRecoveryPointsInput

DescribeTapeRecoveryPointsOutput
DescribeTapeRecoveryPointsOutput

DescribeTapesInput
DescribeTapesInput

DescribeTapesOutput
DescribeTapesOutput

DescribeUploadBufferInput

DescribeUploadBufferOutput

DescribeVTLDevicesInput
DescribeVTLDevicesInput

DescribeVTLDevicesOutput
DescribeVTLDevicesOutput

DescribeWorkingStorageInput
A JSON object containing the of the gateway.

DescribeWorkingStorageOutput
A JSON object containing the following fields:

DeviceiSCSIAttributes
Lists iSCSI information about a VTL device.

DisableGatewayInput
DisableGatewayInput

DisableGatewayOutput
DisableGatewayOutput

Disk

FileShareInfo
Describes a file share.

GatewayInfo
Describes a gateway object.

ListFileSharesInput
ListFileShareInput

ListFileSharesOutput
ListFileShareOutput

ListGatewaysInput

A JSON object containing zero or more of the following fields:

ListGatewaysInput$Limit

ListGatewaysInput$Marker

ListGatewaysOutput

ListLocalDisksInput
A JSON object containing the of the gateway.

ListLocalDisksOutput

ListTagsForResourceInput
ListTagsForResourceInput

ListTagsForResourceOutput
ListTagsForResourceOutput

ListTapesInput

A JSON object that contains one or more of the following fields:

ListTapesInput$Limit

ListTapesInput$Marker

ListTapesInput$TapeARNs

ListTapesOutput

A JSON object containing the following fields:

ListTapesOutput$Marker

ListTapesOutput$VolumeInfos

ListVolumeInitiatorsInput
ListVolumeInitiatorsInput

ListVolumeInitiatorsOutput
ListVolumeInitiatorsOutput

ListVolumeRecoveryPointsInput

ListVolumeRecoveryPointsOutput

ListVolumesInput

A JSON object that contains one or more of the following fields:

ListVolumesInput$Limit

ListVolumesInput$Marker

ListVolumesOutput

NFSFileShareDefaults
Describes file share default values. Files and folders stored as Amazon S3 objects in S3 buckets don't, by default, have Unix file permissions assigned to them. Upon discovery in an S3 bucket by Storage Gateway, the S3 objects that represent files and folders are assigned these default Unix permissions. This operation is only supported in the file gateway architecture.

NFSFileShareInfo
The Unix file permissions and ownership information assigned, by default, to native S3 objects when file gateway discovers them in S3 buckets. This operation is only supported in file gateways.

NetworkInterface
Describes a gateway's network interface.

RefreshCacheInput

RefreshCacheOutput

RemoveTagsFromResourceInput
RemoveTagsFromResourceInput

RemoveTagsFromResourceOutput
RemoveTagsFromResourceOutput

ResetCacheInput

ResetCacheOutput

RetrieveTapeArchiveInput
RetrieveTapeArchiveInput

RetrieveTapeArchiveOutput
RetrieveTapeArchiveOutput

RetrieveTapeRecoveryPointInput
RetrieveTapeRecoveryPointInput

RetrieveTapeRecoveryPointOutput
RetrieveTapeRecoveryPointOutput

SetLocalConsolePasswordInput
SetLocalConsolePasswordInput

SetLocalConsolePasswordOutput

ShutdownGatewayInput
A JSON object containing the of the gateway to shut down.

ShutdownGatewayOutput
A JSON object containing the of the gateway that was shut down.

StartGatewayInput
A JSON object containing the of the gateway to start.

StartGatewayOutput
A JSON object containing the of the gateway that was restarted.

StorageGatewayClient
A client for the AWS Storage Gateway API.

StorageGatewayError
Provides additional information about an error that was returned by the service as an or. See the errorCode and errorDetails members for more information about the error.

StorediSCSIVolume
Describes an iSCSI stored volume.

Tag

Tape
Describes a virtual tape object.

TapeArchive
Represents a virtual tape that is archived in the virtual tape shelf (VTS).

TapeInfo
Describes a virtual tape.

TapeRecoveryPointInfo
Describes a recovery point.

UpdateBandwidthRateLimitInput

A JSON object containing one or more of the following fields:

UpdateBandwidthRateLimitInput$AverageDownloadRateLimitInBitsPerSec

UpdateBandwidthRateLimitInput$AverageUploadRateLimitInBitsPerSec

UpdateBandwidthRateLimitOutput
A JSON object containing the of the gateway whose throttle information was updated.

UpdateChapCredentialsInput

A JSON object containing one or more of the following fields:

UpdateChapCredentialsInput$InitiatorName

UpdateChapCredentialsInput$SecretToAuthenticateInitiator

UpdateChapCredentialsInput$SecretToAuthenticateTarget

UpdateChapCredentialsInput$TargetARN

UpdateChapCredentialsOutput
A JSON object containing the following fields:

UpdateGatewayInformationInput

UpdateGatewayInformationOutput
A JSON object containing the ARN of the gateway that was updated.

UpdateGatewaySoftwareNowInput
A JSON object containing the of the gateway to update.

UpdateGatewaySoftwareNowOutput
A JSON object containing the of the gateway that was updated.

UpdateMaintenanceStartTimeInput

A JSON object containing the following fields:

UpdateMaintenanceStartTimeInput$DayOfWeek

UpdateMaintenanceStartTimeInput$HourOfDay

UpdateMaintenanceStartTimeInput$MinuteOfHour

UpdateMaintenanceStartTimeOutput
A JSON object containing the of the gateway whose maintenance start time is updated.

UpdateNFSFileShareInput
UpdateNFSFileShareInput

UpdateNFSFileShareOutput
UpdateNFSFileShareOutput

UpdateSnapshotScheduleInput

A JSON object containing one or more of the following fields:

UpdateSnapshotScheduleInput$Description

UpdateSnapshotScheduleInput$RecurrenceInHours

UpdateSnapshotScheduleInput$StartAt

UpdateSnapshotScheduleInput$VolumeARN

UpdateSnapshotScheduleOutput
A JSON object containing the of the updated storage volume.

UpdateVTLDeviceTypeInput

UpdateVTLDeviceTypeOutput
UpdateVTLDeviceTypeOutput

VTLDevice
Represents a device object associated with a tape gateway.

VolumeInfo
Describes a storage volume object.

VolumeRecoveryPointInfo

VolumeiSCSIAttributes
Lists iSCSI information about a volume.

Enums

ActivateGatewayError
Errors returned by ActivateGateway

AddCacheError
Errors returned by AddCache

AddTagsToResourceError
Errors returned by AddTagsToResource

AddUploadBufferError
Errors returned by AddUploadBuffer

AddWorkingStorageError
Errors returned by AddWorkingStorage

CancelArchivalError
Errors returned by CancelArchival

CancelRetrievalError
Errors returned by CancelRetrieval

CreateCachediSCSIVolumeError
Errors returned by CreateCachediSCSIVolume

CreateNFSFileShareError
Errors returned by CreateNFSFileShare

CreateSnapshotError
Errors returned by CreateSnapshot

CreateSnapshotFromVolumeRecoveryPointError
Errors returned by CreateSnapshotFromVolumeRecoveryPoint

CreateStorediSCSIVolumeError
Errors returned by CreateStorediSCSIVolume

CreateTapeWithBarcodeError
Errors returned by CreateTapeWithBarcode

CreateTapesError
Errors returned by CreateTapes

DeleteBandwidthRateLimitError
Errors returned by DeleteBandwidthRateLimit

DeleteChapCredentialsError
Errors returned by DeleteChapCredentials

DeleteFileShareError
Errors returned by DeleteFileShare

DeleteGatewayError
Errors returned by DeleteGateway

DeleteSnapshotScheduleError
Errors returned by DeleteSnapshotSchedule

DeleteTapeArchiveError
Errors returned by DeleteTapeArchive

DeleteTapeError
Errors returned by DeleteTape

DeleteVolumeError
Errors returned by DeleteVolume

DescribeBandwidthRateLimitError
Errors returned by DescribeBandwidthRateLimit

DescribeCacheError
Errors returned by DescribeCache

DescribeCachediSCSIVolumesError
Errors returned by DescribeCachediSCSIVolumes

DescribeChapCredentialsError
Errors returned by DescribeChapCredentials

DescribeGatewayInformationError
Errors returned by DescribeGatewayInformation

DescribeMaintenanceStartTimeError
Errors returned by DescribeMaintenanceStartTime

DescribeNFSFileSharesError
Errors returned by DescribeNFSFileShares

DescribeSnapshotScheduleError
Errors returned by DescribeSnapshotSchedule

DescribeStorediSCSIVolumesError
Errors returned by DescribeStorediSCSIVolumes

DescribeTapeArchivesError
Errors returned by DescribeTapeArchives

DescribeTapeRecoveryPointsError
Errors returned by DescribeTapeRecoveryPoints

DescribeTapesError
Errors returned by DescribeTapes

DescribeUploadBufferError
Errors returned by DescribeUploadBuffer

DescribeVTLDevicesError
Errors returned by DescribeVTLDevices

DescribeWorkingStorageError
Errors returned by DescribeWorkingStorage

DisableGatewayError
Errors returned by DisableGateway

ListFileSharesError
Errors returned by ListFileShares

ListGatewaysError
Errors returned by ListGateways

ListLocalDisksError
Errors returned by ListLocalDisks

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTapesError
Errors returned by ListTapes

ListVolumeInitiatorsError
Errors returned by ListVolumeInitiators

ListVolumeRecoveryPointsError
Errors returned by ListVolumeRecoveryPoints

ListVolumesError
Errors returned by ListVolumes

RefreshCacheError
Errors returned by RefreshCache

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

ResetCacheError
Errors returned by ResetCache

RetrieveTapeArchiveError
Errors returned by RetrieveTapeArchive

RetrieveTapeRecoveryPointError
Errors returned by RetrieveTapeRecoveryPoint

SetLocalConsolePasswordError
Errors returned by SetLocalConsolePassword

ShutdownGatewayError
Errors returned by ShutdownGateway

StartGatewayError
Errors returned by StartGateway

UpdateBandwidthRateLimitError
Errors returned by UpdateBandwidthRateLimit

UpdateChapCredentialsError
Errors returned by UpdateChapCredentials

UpdateGatewayInformationError
Errors returned by UpdateGatewayInformation

UpdateGatewaySoftwareNowError
Errors returned by UpdateGatewaySoftwareNow

UpdateMaintenanceStartTimeError
Errors returned by UpdateMaintenanceStartTime

UpdateNFSFileShareError
Errors returned by UpdateNFSFileShare

UpdateSnapshotScheduleError
Errors returned by UpdateSnapshotSchedule

UpdateVTLDeviceTypeError
Errors returned by UpdateVTLDeviceType

Traits

StorageGateway
Trait representing the capabilities of the AWS Storage Gateway API. AWS Storage Gateway clients implement this trait.

Type Definitions

ActivationKey

BandwidthDownloadRateLimit

BandwidthType

BandwidthUploadRateLimit

Boolean

CachediSCSIVolumes

ChapCredentials

ChapSecret

ClientToken

CreatedDate

DayOfWeek

Description

DeviceType

DiskAllocationType

DiskId

DiskIds

Disks

Double

DoubleObject

ErrorCode

ErrorDetails

FileShareARN
The Amazon Resource Name (ARN) of the file share.

FileShareARNList

FileShareClientList
The list of clients that are allowed to access the file gateway. The list must contain either valid IP addresses or valid CIDR blocks.

FileShareId
The ID of the file share.

FileShareInfoList

FileShareStatus
The status of the file share. Possible values are CREATING, UPDATING, AVAILABLE and DELETING.

GatewayARN
The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

GatewayId

GatewayName
The name you configured for your gateway.

GatewayNetworkInterfaces

GatewayOperationalState

GatewayState

GatewayTimezone

GatewayType

Gateways

HourOfDay

IPV4AddressCIDR

Initiator

Initiators

Integer

IqnName

KMSKey
The ARN of the KMS key used for Amazon S3 server side encryption.

LastSoftwareUpdate

LocalConsolePassword

LocationARN
The ARN of the backend storage used for storing file data.

Long

Marker

MediumChangerType

MinuteOfHour

NFSFileShareInfoList

NetworkInterfaceId

NextUpdateAvailabilityDate

NumTapesToCreate

Path
The file share path used by the NFS client to identify the mount point.

PermissionId

PermissionMode

PositiveIntObject

RecurrenceInHours

RegionId

ResourceARN

Role
The ARN of the IAM role that file gateway assumes when it accesses the underlying storage.

SnapshotDescription

SnapshotId

Squash

The user mapped to anonymous user. Valid options are the following:

"RootSquash" - Only root is mapped to anonymous user.

"NoSquash" - No one is mapped to anonymous user

"AllSquash" - Everyone is mapped to anonymous user.

StorageClass

StorediSCSIVolumes

TagKey

TagKeys

TagValue

Tags

TapeARN

TapeARNs
The Amazon Resource Name (ARN) of each of the tapes you want to list. If you don't specify a tape ARN, the response lists all tapes in both your VTL and VTS.

TapeArchiveStatus

TapeArchives

TapeBarcode

TapeBarcodePrefix

TapeDriveType

TapeInfos
An array of TapeInfo objects, where each object describes an a single tape. If there not tapes in the tape library or VTS, then the TapeInfos is an empty array.

TapeRecoveryPointInfos

TapeRecoveryPointStatus

TapeSize

TapeStatus

TapeUsage

Tapes

TargetARN

TargetName

Time

VTLDeviceARN

VTLDeviceARNs

VTLDeviceProductIdentifier

VTLDeviceType

VTLDeviceVendor

VTLDevices

VolumeARN

VolumeARNs

VolumeId

VolumeInfos

VolumeRecoveryPointInfos

VolumeStatus

VolumeType