Struct rusoto_storagegateway::ActivateGatewayInput

pub struct ActivateGatewayInput {
    pub activation_key: String,
    pub gateway_name: String,
    pub gateway_region: String,
    pub gateway_timezone: String,
    pub gateway_type: Option<String>,
    pub medium_changer_type: Option<String>,
    pub tape_drive_type: Option<String>,
}

A JSON object containing one or more of the following fields:

ActivateGatewayInput$ActivationKey

ActivateGatewayInput$GatewayName

ActivateGatewayInput$GatewayRegion

ActivateGatewayInput$GatewayTimezone

ActivateGatewayInput$GatewayType

ActivateGatewayInput$TapeDriveType

ActivateGatewayInput$MediumChangerType

Fields
activation_key: String

Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

For more information, see https://docs.aws.amazon.com/storagegateway/latest/userguide/get-activation-key.html in the Storage Gateway User Guide.

gateway_name: String

The name you configured for your gateway.

gateway_region: String

A value that indicates the region where you want to store your data. The gateway region specified must be the same region as the region in your Host header in the request. For more information about available regions and endpoints for AWS Storage Gateway, see Regions and Endpoints in the Amazon Web Services Glossary.

Valid Values: "us-east-1", "us-east-2", "us-west-1", "us-west-2", "ca-central-1", "eu-west-1", "eu-central-1", "eu-west-2", "eu-west-3", "ap-northeast-1", "ap-northeast-2", "ap-southeast-1", "ap-southeast-2", "ap-south-1", "sa-east-1"

gateway_timezone: String

A value that indicates the time zone you want to set for the gateway. The time zone is of the format "GMT-hr:mm" or "GMT+hr:mm". For example, GMT-4:00 indicates the time is 4 hours behind GMT. GMT+2:00 indicates the time is 2 hours ahead of GMT. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

gateway_type: Option<String>

A value that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is STORED.

Valid Values: "STORED", "CACHED", "VTL", "FILE_S3"

medium_changer_type: Option<String>

The value that indicates the type of medium changer to use for tape gateway. This field is optional.

Valid Values: "STK-L700", "AWS-Gateway-VTL"

tape_drive_type: Option<String>

The value that indicates the type of tape drive to use for tape gateway. This field is optional.

Valid Values: "IBM-ULT3580-TD5"