Struct rusoto_storagegateway::StorediSCSIVolume

pub struct StorediSCSIVolume {
    pub created_date: Option<f64>,
    pub preserved_existing_data: Option<bool>,
    pub source_snapshot_id: Option<String>,
    pub volume_arn: Option<String>,
    pub volume_disk_id: Option<String>,
    pub volume_id: Option<String>,
    pub volume_progress: Option<f64>,
    pub volume_size_in_bytes: Option<i64>,
    pub volume_status: Option<String>,
    pub volume_type: Option<String>,
    pub volume_used_in_bytes: Option<i64>,
    pub volumei_scsi_attributes: Option<VolumeiSCSIAttributes>,
}

Describes an iSCSI stored volume.

Fields
created_date: Option<f64>

The date the volume was created. Volumes created prior to March 28, 2017 don’t have this time stamp.

preserved_existing_data: Option<bool>

Indicates if when the stored volume was created, existing data on the underlying local disk was preserved.

Valid Values: true, false

source_snapshot_id: Option<String>

If the stored volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

volume_arn: Option<String>

The Amazon Resource Name (ARN) of the storage volume.

volume_disk_id: Option<String>

The ID of the local disk that was specified in the CreateStorediSCSIVolume operation.

volume_id: Option<String>

The unique identifier of the volume, e.g. vol-AE4B946D.

volume_progress: Option<f64>

Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the stored volume is not restoring or bootstrapping.

volume_size_in_bytes: Option<i64>

The size of the volume in bytes.

volume_status: Option<String>

One of the VolumeStatus values that indicates the state of the storage volume.

volume_type: Option<String>

One of the VolumeType enumeration values describing the type of the volume.

volume_used_in_bytes: Option<i64>

The size of the data stored on the volume in bytes.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

volumei_scsi_attributes: Option<VolumeiSCSIAttributes>

An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.